🧹 rename go package to packer-plugin-cnspec
