+++ .gitignore
@@ -8,4 +8,5 @@ main
 dist/*
 docs-rendered/*
 packer-plugin-mondoo
+packer-plugin-cnspec
 .docs
