+++ docs/provisioners/mondoo.mdx
@@ -54,4 +54,4 @@ If you are new to Mondoo you can get started by [signing up for a free account](
 Check out the [Building secure AMIs with Mondoo and Packer](https://mondoo.com/docs/tutorials/aws/build-secure-amis-packer/), or [Building secure VM images in Google Cloud with Mondoo and Packer](https://mondoo.com/docs/tutorials/gcp/build-secure-gcp-vm-images-packer/) tutorial on the Mondoo documentation site.
 ## Sample Packer Templates
 
-You can find example Packer templates in the [examples](https://github.com/mondoohq/packer-plugin-mondoo/tree/main/examples) directory in this repository.
+You can find example Packer templates in the [examples](https://github.com/mondoohq/packer-plugin-cnspec/tree/main/examples) directory in this repository.
